move test to correct place
